tasks/crypto/need_more_power/requirements.txt
pycryptodome
fastecdsa
